Updated the scripts to work under many filenames
The lftp scripts will now work and keep separate logs and locks based on the name of the script file. This is handy for people who will be using the script and syncing many separate folders or servers.

I also moved the log files to /var/log as that just makes more sense.

Example:
syncthis.sh would create a `/var/log/syncthis.sh.log` and a `/tmp/syncthis.sh.lock`